+++ ricaun.Nuke.Tests/Tests.cs
@@ -1,0 +1,13 @@
+using NUnit.Framework;
+
+namespace ricaun.Nuke.Tests
+{
+    public class Tests
+    {
+        [Test]
+        public void Test1()
+        {
+            Assert.Pass();
+        }
+    }
+}